big refactoring of the whole project
