+++ .gitignore
@@ -1,1 +1,3 @@
 lodka_experiments/
+experimental_data/
+logs/
