--- utils/2coord-final-err-1m
@@ -1,25 +1,0 @@
-49.898899947694446; 8.901661665633556
-49.898866949330525; 8.901595360886004
-49.898900320761726; 8.901518573191732
-49.898979980242466; 8.901445022069774
-49.89896360557759; 8.901318085341977
-49.89891022610369; 8.901277078573205
-49.89891238020024; 8.901274970451787
-49.89916261563427; 8.901180697885385
-49.89908808456558; 8.900933819917734
-49.89931794895675; 8.900975449917604
-49.89917484754037; 8.900733472393224
-49.899507656609; 8.900845787458689
-49.89925870102962; 8.900533719017927
-49.899645442127465; 8.900702182459087
-49.89931800944349; 8.900328797273842
-49.89975866638337; 8.900646898735895
-49.899357384977456; 8.900120313134174
-49.89977066426223; 8.900441441821465
-49.89939597993179; 8.900074916107227
-49.89973644763797; 8.9002347296525
-49.899610046468254; 8.900085910610269
-49.8997069832671; 8.90009837009444
-49.899742003612076; 8.900061787331738
-49.89976314120955; 8.90001664325857
-49.899818671177414; 8.900040647931725
