+++ utils/devide_area.py
@@ -1,0 +1,121 @@
+#!/usr/bin/env python3
+
+import math
+import numpy as np
+import rospy
+import matplotlib.pyplot as plt
+import os
+import logging, sys
+
+logging.basicConfig(level=logging.INFO)
+boat_number = 4
+
+x = []
+y = []
+# x = [2, 5, 10, 9 ,5]
+# y = [4, 8, 7, 3, 2]
+path = os.path.realpath(__file__)[:-22]
+with open(f'{path}utils/borders') as f:
+    for line in f:
+        line = line.split(';')
+        x.append(float(line[0]))
+        y.append(float(line[1]))
+
+#plt.plot(x, y, 'bo')
+#plt.show()
+
+
+min_lat = min(x)
+max_lat = max(x)
+min_lon = min(y)
+max_lon = max(y)
+len_lat = max_lat - min_lat
+len_lon = max_lon - min_lon
+mid_lat = sum(x)/len(x)
+k_lon = 111320
+k_lat = 111319*math.cos(mid_lat*math.pi/180)
+if len_lat > len_lon:
+    lines = 2
+    cols = boat_number // 2
+    if boat_number % 2 == 1:
+        cols += 1
+else:
+    cols = 2
+    lines = boat_number // 2
+    if boat_number % 2 == 1:
+        lines += 1
+
+areas = [[]]
+for i in range(len(x)):
+    areas[0].append({'lat': x[i], 'lon': y[i]})
+#print(areas[0])
+area_nomber = 0
+
+while area_nomber < boat_number - 1:
+    max_x = -5
+    max_y = -5
+    min_x = 1000
+    min_y = 1000
+    area_nom = len(areas) - 1
+    rospy.logdebug(f'area[nom]_min max {areas[area_nom]}')
+    for coord in areas[area_nom]:
+        rospy.logdebug(f'coord min max {coord}')
+        if coord['lat'] > max_x:
+            max_x = coord['lat']
+        if coord['lon'] > max_y:
+            max_y = coord['lon']
+        if coord['lat'] < min_x:
+            min_x = coord['lat']
+        if coord['lon'] < min_y:
+            min_y = coord['lon']
+        rospy.logdebug(f'min_y {min_y}, max_y {max_y}')
+    len_lat = (max_x - min_x)
+    len_lon = (max_y - min_y)
+    rospy.logdebug(f'len lon {len_lon}')
+    rospy.logdebug(f'next area {areas}')
+    if len_lat/k_lat > len_lon/k_lon:  # Если область широкая делим по вертикали
+        div_lat = min_x + len_lat/2
+        obl1 = []
+        obl2 = []
+        rospy.logdebug(f'area[nom] {areas[area_nom]}')
+        for j in range(len(areas[area_nom])):
+            if areas[area_nom][j]['lat'] < div_lat:
+                obl1.append(areas[area_nom][j])
+            else:
+                obl2.append(areas[area_nom][j])
+            if j == len(areas[area_nom]) - 1:
+                j_next = 0
+            else:
+                j_next = j + 1
+            rospy.logdebug(f'div_lat {div_lat}')
+            rospy.logdebug(f'shir obl1 norm dots: {obl1}')
+            rospy.logdebug(f'shir obl2 norm dots: {obl2}')
+            if ((areas[area_nom][j]['lat'] < div_lat) and (areas[area_nom][j_next]['lat'] > div_lat)) or ((areas[area_nom][j]['lat'] > div_lat) and (areas[area_nom][j_next]['lat'] < div_lat)):
+                a = (areas[area_nom][j]['lon'] - areas[area_nom][j_next]['lon'])/(areas[area_nom][j]['lat'] - areas[area_nom][j_next]['lat'])
+                b = areas[area_nom][j]['lon'] - a*areas[area_nom][j]['lat']
+                lon_per = a*div_lat + b
+                obl1.append({'lat': div_lat, 'lon': lon_per})
+                obl2.append({'lat': div_lat, 'lon': lon_per})
+                rospy.logdebug(f"shir per x:{areas[area_nom][j]['lat']}, y:{areas[area_nom][j]['lon']}, x_next:{areas[area_nom][j_next]['lat']}, y_next:{areas[area_nom][j_next]['lon']}, a: {a}, b: {b}")
+                rospy.logdebug(f'shir lon_per: {lon_per}, div_lat: {div_lat}')
+                rospy.logdebug(f'shir obl1: {obl1}')
+                rospy.logdebug(f'shir obl2: {obl2}')
+        areas.pop(area_nom)
+        areas.insert(0, obl1)
+        areas.insert(0, obl2)
+        area_nomber += 1
+    else:
+        rospy.logdebug(f'area[nom] {areas[area_nom]}')
+        rospy.logdebug(f'dlin min_y {min_y}, max_y {max_y}')
+        rospy.logdebug(f'dlin len lon {len_lon}')
+        div_lon = min_y + len_lon/2
+        obl1 = []
+        obl2 = []
+        for j in range(len(areas[area_nom])):
+            if areas[area_nom][j]['lon'] < div_lon:
+                obl1.append(areas[area_nom][j])
+            else:
+                obl2.append(areas[area_nom][j])
+            if j == len(areas[area_nom]) - 1:
+                j_next = 0
+            else:
