--- utils/way
@@ -1,35 +1,0 @@
-{'lat': 49.8987564727, 'lon': 8.90169695784}
-{'lat': 49.898914397634286, 'lon': 8.9015597651453}
-{'lat': 49.89881918695902, 'lon': 8.901497385692466}
-{'lat': 49.89907232256857, 'lon': 8.901422572450606}
-{'lat': 49.89888190121803, 'lon': 8.901297813544971}
-{'lat': 49.89923024750286, 'lon': 8.901285379755905}
-{'lat': 49.89894461547705, 'lon': 8.901098241397449}
-{'lat': 49.89938817243714, 'lon': 8.901148187061203}
-{'lat': 49.899007329736065, 'lon': 8.900898669249955}
-{'lat': 49.89954609737143, 'lon': 8.901010994366509}
-{'lat': 49.89907004399508, 'lon': 8.900699097102432}
-{'lat': 49.89970402230571, 'lon': 8.900873801671807}
-{'lat': 49.8991327582541, 'lon': 8.90049952495491}
-{'lat': 49.89986194724, 'lon': 8.900736608977105}
-{'lat': 49.899195472513114, 'lon': 8.900299952807416}
-{'lat': 49.899925637, 'lon': 8.90068128035}
-{'lat': 49.89925818677213, 'lon': 8.900100380659893}
-{'lat': 49.8999203868244, 'lon': 8.90047215226491}
-{'lat': 49.899320901031146, 'lon': 8.899900808512399}
-{'lat': 49.899915136648794, 'lon': 8.900263024179822}
-{'lat': 49.89938361529016, 'lon': 8.899701236364876}
-{'lat': 49.8999117653, 'lon': 8.900128734630016}
-{'lat': 49.89944632954918, 'lon': 8.899501664217354}
-{'lat': 49.89990660837607, 'lon': 8.899919604224579}
-{'lat': 49.899509043808195, 'lon': 8.89930209206986}
-{'lat': 49.899901451452145, 'lon': 8.899710473819141}
-{'lat': 49.8995632732, 'lon': 8.899129520849982}
-{'lat': 49.89989629452822, 'lon': 8.899501343413476}
-{'lat': 49.89964448571179, 'lon': 8.89893673431446}
-{'lat': 49.89989113760429, 'lon': 8.899292213008039}
-{'lat': 49.89972569822358, 'lon': 8.89874394777894}
-{'lat': 49.89988598068037, 'lon': 8.899083082602601}
-{'lat': 49.899806910735364, 'lon': 8.898551161243418}
-{'lat': 49.89988082375644, 'lon': 8.898873952196936}
-{'lat': 49.8998692169, 'lon': 8.898403255589997}
